[#92] refactor: listPage, papersPage 에서 사용되는 쿼리훅 파일 위치 변경
